// src/types/KazagumoTypes.ts
export interface KazagumoTrack {
  track: string;
  info: {
    identifier: string;
    title: string;
    author: string;
    length: number;
    artworkUrl?: string;
    uri?: string;
    sourceName: string;
    isSeekable: boolean;
    isStream: boolean;
    position?: number;
  };
  pluginInfo?: any;
  userData?: any;
}